test(browser): separate installer layout coverage
